+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/components/Comment/Comment.js
@@ -1,4 +1,4 @@
-import { Container, Name, Text, Likes } from '../styles/components/CommentStyles'
+import { Container, Name, Text, Likes } from './CommentStyles'
 import { ArrowUpIcon, ArrowDownIcon } from '@chakra-ui/icons'
 
 function Comment() {
