+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/components/Header/Header.js
@@ -1,4 +1,4 @@
-import { Container, Logo } from '../styles/components/HeaderStyles'
+import { Container, Logo } from './HeaderStyles'
 
 function Header() {
     return <Container>
