+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/pages/DetailsPost/DetailsPostStyles.js
