+++ Semana 12 - Revisão/Projeto - LabEddit/labeddit/src/pages/Login/LoginPage.js
@@ -1,24 +1,20 @@
 import { useState } from 'react'
-import { useHistory } from 'react-router-dom'
-import { goToLogin } from '../router/coordinator'
+import useForm from '../../hooks/useForm'
 import { InputGroup, InputRightElement } from '@chakra-ui/react'
-import { InputPattern } from "../components/InputPattern"
-import { ButtonPattern } from '../components/ButtonPattern'
-import { Container, Card, Title, Text, ButtonRegister } from '../styles/pages/RegisterStyles'
+import { InputPattern } from "../../components/InputPattern"
+import { ButtonPattern } from '../../components/ButtonPattern'
+import { Container, Card, Title, Text, ButtonRegister } from './LoginStyles'
 import { ViewIcon, ViewOffIcon } from '@chakra-ui/icons'
-import { Button, ContainerButton } from '../styles/ButtonFormStyles'
-import Footer from '../components/Footer'
-import useForm from '../hooks/useForm'
+import { Button, ContainerButton } from '../../styles/ButtonFormStyles'
+import Footer from '../../components/Footer/Footer'
 
-function RegisterPage() {
+function LoginPage() {
   const [show, setShow] = useState(false)
   const handleClick = () => setShow(!show)
-  const history = useHistory()
 
   const initialState = {
     email: '',
     password: '',
-    username: ''
   }
   const [form, onChange, resetForm] = useForm(initialState)
 
@@ -26,30 +22,22 @@ function RegisterPage() {
     event.preventDefault()
     resetForm()
   }
-
   return <Container>
     <Card>
-      <Title>Cadastro</Title>
-
+      <Title>Login</Title>
       <form onSubmit={onSubmitForm}>
         <InputPattern
-          placeholder="Nome"
-          type={'text'}
-          name={'username'}
-          value={form.username}
-          onChange={onChange}
-        />
-        <InputPattern
-          placeholder="Email"
-          type={'email'}
+          placeholder={'Email'}
           name={'email'}
           value={form.email}
           onChange={onChange}
+          type={'email'}
+          pattern={'[^@ \t\r\n]+@[^@ \t\r\n]+\\.[^@ \t\r\n]+'}
         />
 
-        <InputGroup>
+        <InputGroup size="md">
           <InputPattern
-            placeholder="Senha"
+            placeholder={"Senha"}
             name={'password'}
             value={form.password}
             onChange={onChange}
@@ -70,10 +58,10 @@ function RegisterPage() {
         </InputGroup>
 
         <ContainerButton>
-          <Button type={'submit'}>Cadastrar</Button>
+          <Button type={'submit'}>Entrar</Button>
         </ContainerButton>
       </form>
-      <Text>Já tem uma conta? Faça <ButtonRegister onCLick={() => goToLogin(history)}>login</ButtonRegister> agora</Text>
+      <Text>Não tem uma conta? Se <ButtonRegister>cadastre</ButtonRegister> agora</Text>
     </Card>
     <Footer
       background={'rgb(0, 0, 0)'}
@@ -82,4 +70,4 @@ function RegisterPage() {
   </Container>
 }
 
-export default RegisterPage
+export default LoginPage
